Create example distro squashfs (from build host)
 - Add a target 'example-distro' which must be manually invoked
   This creates an example.sqfs with a minimal binary distro
   based on the build hosts's /bin and /lib.

The purpose is to provide a simple test image which the
default configuration of the unpackfs module can use to
create a system within which the *other* steps of the
installation can run.

Example files are some zoneinfo's (remember to choose an
existing zone when using the example distro), groups and
sudoers files, etc ..

The example distro has a special /xbin which contains bogus
binaries for many system-administration tasks (e.g. useradd
which would otherwise come from /usr/sbin).