[netinstall] Expand the table of common-strings
When one of these common names for the netinstall page is used,
it gets pulled out of the standard translations, so that it
doesn't have to be translated in the per-distro config file.
These labels are common enough that they make sense for
everyone to have lying around.

FIXES #1367

(I say "fixed" but of course it's going to depend on the translation
workflow to make these available)