CI: update clang-format
In clang-format 10, SpaceInEmptyBlock is introduced, and defaults to
true .. which is different from the earlier formatting versions did.
For now, refuse clang-format 10, and search specifically also for
clang-format-9.0.1 because that's what I have on my laptop.

At some point, switch in the config option and then require
clang-format 10 or later (because earlier versions refuse to
run with an unknown config option)