[partition] Repair translation of labels
There was an accumulation of connected slots when the
current device changed (CALAMARES_RETRANSLATE **adds** one).
Instead, extract the translated messages and carry the
necessary information (number of entries, and if just one,
its name) in the ChoicePage object.

Although you can't effectively change the translation /
current language while looking at the ChoicePage, it is
best to be future-proof.