[netinstall] Remove unnecessary $id from schema
The base-URI of the schema is already set; references
that are relative and contain only a fragment (e.g. start with `#`)
resolve against the base-URI, so that's within this document.
Since we already have `definitions` as a key, `#/definitions`
references that sub-schema. We don't need the `$id` settings in
each sub-schema, which messes up the base-URI; it should have been
an `$anchor`, maybe, but isn't necessary after all.