[libcalamares] Block sleep with systemd/CK login manager if present
Calamares running as root is not always conducive to successfully
communicating with the fd.o PowerManagement interface on the user
session bus.

The login manager interface has long provided a similarly capable
sleep inhibition functionality, which at this point would be present
on the vast majority of distributions. Use it if systemd-logind
or ConsoleKit2 have registered this service on the system bus.
Regardless of running as root or user, Calamares shouldn't have
an issue contacting this interface.

Fixes #2384